moving model_types to own submodule to readd imports without circular dependicies
