+++ evalml/__init__.py
@@ -12,7 +12,10 @@ import evalml.objectives
 import evalml.tuners
 import evalml.demos
 import evalml.problem_types
+import evalml.pipelines
+import evalml.model_types
 
+from evalml.pipelines import list_model_types, save_pipeline, load_pipeline
 from evalml.models import AutoClassifier, AutoRegressor
 
 __version__ = '0.4.1'
