+++ evalml/model_types/__init__.py
@@ -1,0 +1,2 @@
+# flake8:noqas
+from .model_types import ModelTypes
