+++ evalml/pipelines/classification/logistic_regression.py
@@ -7,7 +7,7 @@ from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import StandardScaler
 from skopt.space import Real
 
-from evalml.models.model_types import ModelTypes
+from evalml.model_types import ModelTypes
 from evalml.pipelines import PipelineBase
 from evalml.problem_types import ProblemTypes
