+++ evalml/pipelines/classification/xgboost.py
@@ -7,7 +7,7 @@ from sklearn.pipeline import Pipeline
 from skopt.space import Integer, Real
 from xgboost import XGBClassifier
 
-from evalml.models.model_types import ModelTypes
+from evalml.model_types import ModelTypes
 from evalml.pipelines import PipelineBase
 from evalml.problem_types import ProblemTypes
